vendor and client side completed
client can now book vendors and select packages, when the order is placed, the total bill is displayed. The user order dashboard is updated with the name of the chosen vendor and the status for the order is displayed as well. On the vendor side, the vendor can accept/reject an order. when the vendor accepts/rejects an order the vendor dashboard is updated and the pending orders are removed from the dashboard. On the client side, the status of the orders are updated